finish execution logs delegate
